Factor generic flag machinery into utils _flags module
